cleaning, improving and updating code
