+++ register.h
@@ -1,6 +1,6 @@
 /*
 	This library contains a definition of the
-	data type of the "reg" and methods to 
+	data type of the "reg" and methods to
 	manipulate dinamic lists of him
 */
 
@@ -10,14 +10,14 @@
 
 //Dinamic data struct for lists
 typedef struct regType{
-	struct regType *prev; 
+	struct regType *prev;
 	struct regType *next;
-	int id; 
+	int id;
 	char description[100];
-	char category[50];	
-	char details[100]; 
-	char date[20];		
-	double value;		
+	char category[50];
+	char details[100];
+	char date[20];
+	double value;
 }reg;
 
 //Function to initalize strings
@@ -29,7 +29,7 @@ void initalizeString(char* String, int Size){
 }
 
 //Create an empty register
-reg *newReg(){
+reg *newReg(void){
     reg *emptyReg = (reg *)malloc(sizeof(reg));						//Allocates *reg size memory
     emptyReg->next = emptyReg->prev = NULL;							//Initalize poiters with NULL
     initalizeString(emptyReg->date,20);								//Initialize date with \0
@@ -44,12 +44,12 @@ reg *newReg(){
 //Returns to the end of the list
 reg *navEnd(reg *regsList){
 	return (regsList->next != NULL ? navEnd(regsList->next) : regsList);
-};
+}
 
 //Returns to the beginning of the list
 reg *navStart(reg *regsList){
 	return (regsList->prev != NULL ? navStart(regsList->prev) : regsList);
-};
+}
 
 //Add a new item at the beginning of the list
 reg *addStart(reg *regsList, reg *newItem){
@@ -58,7 +58,7 @@ reg *addStart(reg *regsList, reg *newItem){
 	regsList->prev = newItem;
 	newItem->next = regsList;
 	return newItem;
-};
+}
 
 //Add a new item at the middle of the list
 reg *addMiddle(reg *regsList, reg *newItem){
@@ -67,7 +67,7 @@ reg *addMiddle(reg *regsList, reg *newItem){
 	newItem->prev = regsList;
 	regsList->next = newItem;
 	return navStart(regsList);
-};
+}
 
 //Add a new item to the end of the list
 void addEnd(reg *regsList, reg *newItem){
@@ -75,7 +75,7 @@ void addEnd(reg *regsList, reg *newItem){
 	newItem->prev = regsList;
 	newItem->next = regsList->next;
 	regsList->next = newItem;
-};
+}
 
 //Remove the first item at the beginning of the list
 reg *remStart(reg *regsList){
@@ -84,7 +84,7 @@ reg *remStart(reg *regsList){
 	(temp != NULL) ? (temp->prev = regsList->prev) : 0;
 	free(regsList);
 	return temp;
-};
+}
 
 //Remove an item in the middle of the list
 reg *remMiddle(reg *regsList){
@@ -95,11 +95,24 @@ reg *remMiddle(reg *regsList){
 		free(regsList);
 	}
 	return temp;
-};
+}
 
 //Remove the last item at the end of the list
 void remEnd(reg *regsList){
 	regsList = navEnd(regsList);
 	(regsList->prev != NULL) ? (regsList->prev->next = regsList->next) : 0;
 	free(regsList);
-};
+}
+
+//Function to clean list
+void cleanList(reg *regList){
+	
+	//Verifying if regList is null
+	if(regList != NULL && regList->next != NULL){
+			
+		//Function call to remove records
+		remEnd(regList);
+		cleanList(regList);
+		
+	}
+}
